Add Envoy Gateway as ingress mechanism for integration tests
This change integrates _Envoy Gateway_ into the integration test
environment to handle the ingress traffic for the Kind cluster. The
implementation uses _Helm_ to install _Envoy Gateway_ from an OCI
registry, configures it to use `NodePort` service type, and sets up
default gateway resources.

A side effect of this change is that the traffic will be redirected from
`127.0.0.1:8000` to their destination using TLS SNI, so according to the
destination host name. So `127.0.0.1:8000` will not work, it will be
necessary to use the same names that are used inside the cluster. For
example, for the API the host name will now be
`fulfillment-api.innabox.svc.cluster.local`.

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>