Add home directory support for shorter log messages
Added a `SetHome` method to the `CommandBuilder` and `KindBuilder` that
allows setting a project home directory. When set, log messages display
subdirectories relative to the home directory using '~' prefix, making
them more readable.

The `Command` builder now shortens directory paths in log messages by
replacing the home directory prefix with '~' when the directory is a
subdirectory of the home. This is applied to all command executions in
the `Kind` cluster helper.

Changes include:

- Added `SetHome` method to `CommandBuilder`.
- Added `home` field to `CommandBuilder` and `Command` types.
- Modified logger attribute creation to shorten directory paths.
- Added `SetHome` method to `KindBuilder`.
- Added `home` field to `KindBuilder` and `Kind` types.
- Updated all command creation calls in `Kind` to pass home directory.

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>