Add `LocalhostCertificateFiles` testing function
This adds a new `LocalhostCertificateFiles` function that writes the
certificate, key, and CA returned by `LocalhostCertificate` to temporary
files. The files are PEM encoded. Since the certificate is self-signed,
the CA file contains the same certificate.

The caller is responsible for deleting these temporary files when they
are no longer needed.

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>